Refactor terminal components to no longer need api origin prop
