+++ packages/webcomponents/src/api/services/terminal_orders.service.ts
@@ -1,11 +1,13 @@
-import { Api, IApiResponseCollection, ITerminalOrder } from '..';
+import { IApiResponseCollection, ITerminalOrder } from '..';
+import Api from '../ApiNew';
+
+const api = Api();
 
 export interface ITerminalOrderService {
   fetchTerminalOrders(
     accountId: string,
     authToken: string,
-    params: any,
-    apiOrigin?: string
+    params: any
   ): Promise<IApiResponseCollection<ITerminalOrder[]>>;
 }
 
@@ -13,17 +15,10 @@ export class TerminalOrderService implements ITerminalOrderService {
   async fetchTerminalOrders(
     accountId: string,
     authToken: string,
-    params: any,
-    apiOrigin: string = PROXY_API_ORIGIN
+    params: any
   ): Promise<IApiResponseCollection<ITerminalOrder[]>> {
-    const headers = { account: accountId };
-
-    const api = Api({ authToken, apiOrigin });
     const endpoint = 'terminals/orders';
-    return api.get({
-      endpoint,
-      params,
-      headers,
-    });
+    const headers = { account: accountId };
+    return api.get({ endpoint, params, headers, authToken });
   }
 }
